+++ includes/class-hrs-theme-setup.php
@@ -45,7 +45,7 @@ class HRS_Theme_Setup {
 	}
 
 	/**
-	 * Loads the WP API actions and hooks.
+	 * Loads the WP API and Spine actions and hooks.
 	 *
 	 * @since 0.12.0
 	 *
@@ -54,7 +54,46 @@ class HRS_Theme_Setup {
 	private function setup_hooks() {
 		// add theme support
 		// register nav menu
-		// maybe set Spine defaults using its hook?
+
+		// Set Spine options.
+		add_action( 'after_setup_theme', array( $this, 'get_hrs_spine_schema' ), 5 );
+		add_filter( 'spine_enable_builder_module', '__return_true' );
+	}
+
+	/**
+	 * Returns the Spine version for this site.
+	 *
+	 * Retrieves the Spine version if it exists, and calls the setter if it
+	 * doesn't. The Spine schema version is used by the Spine parent theme to
+	 * provide the most current set of default options for the Spine config.
+	 *
+	 * @since 0.12.0
+	 *
+	 * @return string The schema version number
+	 */
+	public function get_hrs_spine_schema() {
+		$hrs_schema = get_option( 'spine_schema', '' );
+
+		if ( '' === $hrs_schema ) {
+			$hrs_schema = self::set_hrs_spine_schema();
+		}
+
+		return $hrs_schema;
+	}
+
+	/**
+	 * Sets the Spine schema version for this site.
+	 *
+	 * @since 0.12.0
+	 *
+	 * @return string The schema version number
+	 */
+	public function set_hrs_spine_schema() {
+		$hrs_schema = '2.x';
+
+		update_option( 'spine_schema', $hrs_schema );
+
+		return $hrs_schema;
 	}
 
 }